autocrop working, but requires openCV 4+
autocrop is now working again, but requires openCV4+ as the call out to findContours() returns a different number of variables. All my systems will be upgraded and I see no reason to support the older versions at this time.

I added a call out to the parent image to get the max size of the image-to-be-cropped to correctly identify the crop coordinates in Capture One Pro. It should now work regardless of the orignal camera used -- before height and width for setting dimensions was hard-coded in with the Fuji 50s defaults.

This call out should probably be moved into a method and only called if necessary as it seems to have slowed things down a bit. If I can get support and time to refactor my autocrop into coding4ch this would be a priority.